requirements.txt
numpy
scikit-image
scikit-learn
pillow
torch
torchvision